Added bullets and particles

+++ EngineMETA/TileMap.h
@@ -4,13 +4,13 @@
 
 class TileMap {
   public:
-    uint8_t *data;
+    int *data;
     void draw();
-    uint8_t tileWidth, tileHeight; //the size of the tiles in number of pixels
-    uint8_t widthTiles, heightTiles; //the size of the map in number of tiles
+    int tileWidth, tileHeight; //the size of the tiles in number of pixels
+    int widthTiles, heightTiles; //the size of the map in number of tiles
 
     TileMap();
-    int16_t getTile(int16_t x, int16_t y);
+    int getTile(int x, int y);
 };
 
 #endif // _TILEMAP_H_
