+++ EngineMETA/Toolbox.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 float norm(float n);
-bool collidePointRect(int16_t x1, int16_t y1 ,int16_t x2 ,int16_t y2, int16_t w, int16_t h);
-bool collideRectRect(int16_t x1, int16_t y1, int16_t w1, int16_t h1 ,int16_t x2 ,int16_t y2, int16_t w2, int16_t h2);
+bool collidePointRect(int x1, int y1 ,int x2 ,int y2, int w, int h);
+bool collideRectRect(int x1, int y1, int w1, int h1 ,int x2 ,int y2, int w2, int h2);
 
 #endif // _TOOLBOX_H_
